backend/requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-filter==24.3
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
dj-database-url==2.2.0
redis==5.1.1
django-redis==5.4.0
Pillow==11.0.0
python-magic==0.4.27
boto3==1.35.53
django-storages==1.14.4
celery==5.4.0
django-celery-beat==2.7.0
openai==1.68.2
python-dateutil==2.9.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
sentry-sdk[django]==2.17.0
